@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev -p 3000",
+    "dev": "NEXT_PRIVATE_LOCAL_WEBPACK=true next dev -p 3000",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
@@ -12,21 +12,23 @@
     "test:report": "playwright show-report"
   },
   "dependencies": {
+    "@healthcare-portal/shared-library": "*",
     "@module-federation/nextjs-mf": "^8.8.4",
-    "next": "14.1.0",
+    "next": "^14.2.28",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "@healthcare-portal/shared-library": "*",
-    "webpack": "^5.99.7"
+    "winston": "^3.11.0"
   },
   "devDependencies": {
     "@playwright/test": "^1.42.1",
     "@types/node": "^20.0.0",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
+    "@types/winston": "^2.4.4",
     "autoprefixer": "^10.4.21",
     "postcss": "^8.5.3",
     "tailwindcss": "^3.4.17",
-    "typescript": "^5.0.0"
+    "typescript": "^5.0.0",
+    "webpack": "^5.99.7"
   }
-} 
+}
