@@ -9,11 +9,11 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@healthcare-portal/shared-library": "*",
     "@module-federation/nextjs-mf": "^8.8.4",
-    "next": "14.1.0",
+    "next": "^14.2.28",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "@healthcare-portal/shared-library": "*"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@types/node": "^20.0.0",
@@ -21,4 +21,4 @@
     "@types/react-dom": "^18.2.0",
     "typescript": "^5.0.0"
   }
-} 
+}
