@@ -1,10 +1,10 @@
 import type { AppProps } from 'next/app';
-import { AppProvider } from '@healthcare-portal/shared-library';
+import { AuthProvider } from '@healthcare-portal/shared-library';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
-    <AppProvider>
+    <AuthProvider value={{ user: null, isLoading: false, error: null }}>
       <Component {...pageProps} />
-    </AppProvider>
+    </AuthProvider>
   );
 } 
